requirements.txt
Flask==2.2.5
gunicorn==21.2.0
requests==2.32.5
python-dotenv==1.0.1
# Opcionales (FASE 2)
openai==1.30.1
httpx==0.28.1
gspread==6.0.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
